requirements.txt
pre-commit
pytest
django
pymysql
djangorestframework
mysqlclient
psycopg2-binary